+++ pysat/_instrument.py
@@ -1975,7 +1975,7 @@ class Instrument(object):
         return export_dict
 
     def to_netcdf4(self, fname=None, base_instrument=None, epoch_name='Epoch',
-                   zlib=False, complevel=4, shuffle=True):
+                   zlib=False, complevel=4, shuffle=True, preserve_meta_case=False):
         """Stores loaded data into a netCDF4 file.
 
         Parameters
@@ -1996,6 +1996,13 @@ class Instrument(object):
             the HDF5 shuffle filter will be applied before compressing the data
             (default True). This significantly improves compression. Default is
             True. Ignored if zlib=False.
+        preserve_meta_case : bool (False)
+            if True, then the variable strings within the MetaData object, which
+            preserves case, are used to name variables in the written netCDF file.
+            If False, then the variable strings used to access data from the
+            Instrument object are used instead. By default, the variable strings
+            on both the data and metadata side are the same, though this relationship
+            may be altered by a user.
 
         Note
         ----
@@ -2013,7 +2020,12 @@ class Instrument(object):
            structure
 
 
-        All attributes attached to instrument meta are written to netCDF attrs.
+        All attributes attached to instrument meta are written to netCDF attrs
+        with the exception of 'Date_End', 'Date_Start', 'File', 'File_Date',
+        'Generation_Date', and 'Logical_File_ID'. These are defined within to_netCDF
+        at the time the file is written, as per the adopted standard, 
+        SPDF ISTP/IACG Modified for NetCDF. Atrributes 'Conventions' and 
+        'Text_Supplement' are given default values if not present.
 
         """
 
@@ -2049,6 +2061,16 @@ class Instrument(object):
         # Apply instrument specific post-processing to the export_meta
         if hasattr(self._export_meta_post_processing, '__call__'):
             export_meta = self._export_meta_post_processing(export_meta)
+
+        # check if there are multiple variables with same characters
+        # but with different case
+        variables = self.variables
+        variables = [var.lower() for var in variables]
+        unique_variables = np.unique(variables)
+        if len(unique_variables) != len(variables):
+            raise ValueError('There are multiple variables with the same ' +
+                             'name but different case which results in a ' +
+                             'loss of metadata. Please make the names unique.')
 
         # general process for writing data is this
         # first, take care of the EPOCH information
@@ -2120,13 +2142,20 @@ class Instrument(object):
                 # coltype is the direct type, np.int64
                 # and datetime_flag lets you know if the data is full of time
                 # information
+                if preserve_meta_case:
+                    # use the variable case stored in the MetaData object
+                    case_key = self.meta.var_case_name(key)
+                else:
+                    # use variable names used by user when working with data
+                    case_key = key
                 data, coltype, datetime_flag = self._get_data_info(self[key],
                                                                    file_format)
                 # operate on data based upon type
                 if self[key].dtype != np.dtype('O'):
                     # not an object, normal basic 1D data
                     # print(key, coltype, file_format)
-                    cdfkey = out_data.createVariable(key,
+
+                    cdfkey = out_data.createVariable(case_key,
                                                      coltype,
                                                      dimensions=(epoch_name),
                                                      zlib=zlib,
@@ -2135,7 +2164,7 @@ class Instrument(object):
                     # attach any meta data, after filtering for standards
                     try:
                         # attach dimension metadata
-                        new_dict = export_meta[key]
+                        new_dict = export_meta[case_key]
                         new_dict['Depend_0'] = epoch_name
                         new_dict['Display_Type'] = 'Time Series'
                         new_dict['Format'] = self._get_var_type_code(coltype)
@@ -2167,7 +2196,8 @@ class Instrument(object):
 
                     if (coltype == type(' ')) or (coltype == type(u' ')):
                         # dealing with a string
-                        cdfkey = out_data.createVariable(key, coltype,
+                        cdfkey = out_data.createVariable(case_key,
+                                                         coltype,
                                                          dimensions=(epoch_name),
                                                          zlib=zlib,
                                                          complevel=complevel,
@@ -2175,7 +2205,7 @@ class Instrument(object):
                         # attach any meta data
                         try:
                             # attach dimension metadata
-                            new_dict = export_meta[key]
+                            new_dict = export_meta[case_key]
                             new_dict['Depend_0'] = epoch_name
                             new_dict['Display_Type'] = 'Time Series'
                             new_dict['Format'] = \
@@ -2213,7 +2243,7 @@ class Instrument(object):
                         for i, dim in enumerate(dims[:-1]):
                             # don't need to go over last dimension value,
                             # it covers number of columns (if a frame)
-                            obj_dim_names.append(key)
+                            obj_dim_names.append(case_key)
                             out_data.createDimension(obj_dim_names[-1], dim)
                         # create simple tuple with information needed to create
                         # the right dimensions for variables that will
@@ -2257,7 +2287,7 @@ class Instrument(object):
                                 data, coltype, _ = \
                                     self._get_data_info(idx, file_format)
                                 cdfkey = \
-                                    out_data.createVariable(key+'_'+col,
+                                    out_data.createVariable(case_key + '_' + col,
                                                             coltype,
                                                             dimensions=var_dim,
                                                             zlib=zlib,
@@ -2265,7 +2295,7 @@ class Instrument(object):
                                                             shuffle=shuffle)
                                 # attach any meta data
                                 try:
-                                    new_dict = export_meta[key+'_'+col]
+                                    new_dict = export_meta[case_key + '_'+col]
                                     new_dict['Depend_0'] = epoch_name
                                     new_dict['Depend_1'] = obj_dim_names[-1]
                                     new_dict['Display_Type'] = 'Spectrogram'
@@ -2302,7 +2332,7 @@ class Instrument(object):
                                 data, coltype, _ = \
                                     self._get_data_info(idx, file_format)
                                 cdfkey = \
-                                    out_data.createVariable(key + '_data',
+                                    out_data.createVariable(case_key + '_data',
                                                             coltype,
                                                             dimensions=var_dim,
                                                             zlib=zlib,
@@ -2310,7 +2340,7 @@ class Instrument(object):
                                                             shuffle=shuffle)
                                 # attach any meta data
                                 try:
-                                    new_dict = export_meta[key]
+                                    new_dict = export_meta[case_key]
                                     new_dict['Depend_0'] = epoch_name
                                     new_dict['Depend_1'] = obj_dim_names[-1]
                                     new_dict['Display_Type'] = 'Spectrogram'
@@ -2345,13 +2375,13 @@ class Instrument(object):
                                                 file_format)
                         # create dimension variable for to store index in
                         # netCDF4
-                        cdfkey = out_data.createVariable(key, coltype,
+                        cdfkey = out_data.createVariable(case_key, coltype,
                                                          dimensions=var_dim,
                                                          zlib=zlib,
                                                          complevel=complevel,
                                                          shuffle=shuffle)
                         # work with metadata
-                        new_dict = export_meta[key]
+                        new_dict = export_meta[case_key]
                         new_dict['Depend_0'] = epoch_name
                         new_dict['Depend_1'] = obj_dim_names[-1]
                         new_dict['Display_Type'] = 'Time Series'
@@ -2415,28 +2445,44 @@ class Instrument(object):
                 if key not in base_attrb:
                     if key[0] != '_':
                         adict[key] = self.meta.__getattribute__(key)
+            # Add additional metadata to conform to standards
             adict['pysat_version'] = pysat.__version__
             if 'Conventions' not in adict:
                 adict['Conventions'] = 'SPDF ISTP/IACG Modified for NetCDF'
             if 'Text_Supplement' not in adict:
-                adict['Text_Supplement'] = ''
+                adict['Text_Supplement'] = ''                
+            # remove any attributes with the names below
+            # pysat is responible for including them in the file.
+            items = ['Date_End', 'Date_Start', 'File', 'File_Date',
+                     'Generation_Date', 'Logical_File_ID']
+            for item in items:
+                if item in adict:
+                    _ = adict.pop(item)
 
-            adict['Date_Start'] = \
-                pysat.datetime.strftime(self.index[0],
-                                        '%a, %d %b %Y,  ' +
-                                        '%Y-%m-%dT%H:%M:%S.%f UTC')
             adict['Date_End'] = \
                 pysat.datetime.strftime(self.index[-1],
                                         '%a, %d %b %Y,  ' +
-                                        '%Y-%m-%dT%H:%M:%S.%f UTC')
+                                        '%Y-%m-%dT%H:%M:%S.%f')
+            adict['Date_End'] = adict['Date_End'][:-3] + ' UTC'
+            
+            adict['Date_Start'] = \
+                pysat.datetime.strftime(self.index[0],
+                                        '%a, %d %b %Y,  ' +
+                                        '%Y-%m-%dT%H:%M:%S.%f')
+            adict['Date_Start'] = adict['Date_Start'][:-3] + ' UTC'
             adict['File'] = os.path.split(fname)
+            adict['File_Date'] = \
+                self.index[-1].strftime('%a, %d %b %Y,  ' +
+                                        '%Y-%m-%dT%H:%M:%S.%f')
+            adict['File_Date'] = adict['File_Date'][:-3] + ' UTC'
             adict['Generation_Date'] = \
                 pysat.datetime.utcnow().strftime('%Y%m%d')
             adict['Logical_File_ID'] = os.path.split(fname)[-1].split('.')[:-1]
-            # check for binary types
+
+            # check for binary types, convert when found
             for key in adict.keys():
                 if isinstance(adict[key], bool):
                     adict[key] = int(adict[key])
-            # print('adict', adict)
+            # attach attributes
             out_data.setncatts(adict)
         return